Fix truncated CGI responses
CGI responses would be randomly truncated at most page loads, but not
always. That's because most CGI responses were long enough/slow enough
to be written in multiple stdoutResponseWriter.Write() calls. Each
call would also write its own copy of the HTTP, messing up the
response. I also think the headers got an incorrect Content-Length
that covered only one of the writes, not the whole page.

stdoutResponseWriter must conform to an interface that uses value
receivers instead of pointer receivers. Since we have a value not a
pointer, we can't modify the struct in the methods. (This means that
every non-200 statusCode was probably lost too by the old code.) So
now we just use a dummy struct with no fields so that we conform to
the interface, store the actual CGI state in global variables, and
write it all out at the end.